requirements.txt
aioredis==2.0.1
anyio==4.9.0
async-timeout==5.0.1
fastapi==0.115.12
httptools==0.6.4
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
starlette==0.46.1
uvicorn==0.34.0
websockets==15.0.1